requirements.txt
streamlit
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index
Pillow
python-dotenv
google-generativeai